Move sizeToFitLayers(_:) function from SBLayoutHandler.swift to SBStarRatingControl.swift and delete SBLayoutHandler.swift file
